SQLite has been integrated to the project.
Now all the session data is stored in a SQLite database. Profiles
are introduced to the project which are just JSON files used
to store tokens and address of the server. Some refactoring, and minor
improvements have been made to the code.

We are going to divide and organize the single server file more
so code is clean and easy to maintain.