Removed Admin user type.
There is no more separate admin user support. the admin console is going
to be embedded in the server file. That means the one that initiates the
server will have the console right away.

Also set the 'connect-ip' flag to localhost for the tests. To solve the
bug that occurs when testing