Bug fixes, and other improvements.
This commit introduces a clarity fix for the command_status_code flag.
Previously if you wanted to show the end of a command execution you would
send a status code through the 'command_status_code' property from the
victim. but now the response itself is a dict and the command_status_code
is separate.