Added dependency containers.
Using the new dependency containers, the run_server function now found
in server/main will use the container instance to get all the dependencies
of the server. There are two implementations of the server container one
for the tests and one for the default execution.

The database is also now a separate dependency instead of being initialized
from the config. The config only provides the path for the db.

This should help future maintainability of the code base.